backend/agents/requirements.txt
langgraph==0.0.20
anthropic==0.18.1
google-generativeai==0.3.2
openai==1.12.0
requests==2.31.0
docker==7.0.0
langgraph==0.0.40
langchain==0.1.0
anthropic==0.18.0
boto3==1.34.0
PyGithub==2.1.1
GitPython==3.1.41
boto3==1.34.34
requests==2.31.0
